refactor: Move 'extractToken', 'signJWS', and 'verifyToken' to 'util'
