+++ api/modules/middleware/has-jwt.ts
@@ -1,47 +1,9 @@
 import type { NextFunction, Request, Response } from 'express';
-import * as jose from 'jose';
 
-import config from '../../config';
 import AppError from '../../util/app-error';
+import { extractToken, verifyToken } from '../../util/header-and-jwt';
 import CacheService from '../cache/service';
 import UserService from '../user/service';
-
-/**
- * This function is used to verify the current token from user's cookie.
- * We will also check the supposed 'audience' and 'issuer'.
- *
- * @param token - JWT token
- * @returns A promise object that contains our JWT.
- */
-const verifyToken = async (token: string) => {
-  const publicKey = await jose.importSPKI(config.JWT_PUBLIC_KEY, 'EdDSA');
-
-  const options: jose.JWTVerifyOptions = {
-    audience: config.JWT_AUDIENCE,
-    issuer: config.JWT_ISSUER,
-  };
-
-  return jose.jwtVerify(token, publicKey, options);
-};
-
-/**
- * Extracts a token from either Authorization header or signed cookie.
- * Prioritize token from Authorization header.
- *
- * @param header - Authorization header value
- * @param signedCookie - Signed cookie if applicable
- * @returns Extracted JWT token
- */
-const extractToken = (
-  header: string | undefined,
-  signedCookie: string | undefined
-) => {
-  if (header && header.startsWith('Bearer ')) {
-    return header.split(' ')[1];
-  }
-
-  return signedCookie;
-};
 
 /**
  * Validates whether a user is authorized/authenticated or not (via JWT).
